Move ActionStepBlock into Actions directory.
