+++ resources/assets/components/Actions/index.js
@@ -1,3 +1,4 @@
+export * from './ActionStepBlock';
 export * from './LinkAction';
 export * from './ThirdPartyAction';
 export * from './VoterRegistrationAction';
