+++ resources/assets/components/Page/ActionPage/index.js
@@ -1,10 +1,5 @@
 // Fully constructed Action Page.
 export ActionPageContainer from './ActionPageContainer';
 
-// Fully constructed Action Steps Wrapper.
-export ActionStepsWrapperContainer from './ActionStepsWrapperContainer';
-
-export ActionStepsWrapper from './ActionStepsWrapper';
-
 // Pure React component without Redux data.
 export default from './ActionPage';
